requirements.txt
Django>=4.3
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
drf-yasg==1.21.7
django-cors-headers
pillow
python-magic
dj_database_url
gunicorn
openai
httpx
google-genai
ruff
black
